Mod03/requirements.txt
streamlit 
langchain
langchain-community
langchain-openai
openai
chromadb
pypdf
python-dotenv
tiktoken
